nest: use strings joins rather than +=, re #111
Append to a list of strings, rather than concatenating to one string very often or printing often.
We only print at the end.

This is done in most of the actual map generation and saves
ca 0.1 seconds.
The map_generation itself now takes ca. 0.1 seconds.

The biggest part is now 0.5 seconds for the database fetch.


git-svn-id: http://svn.kraehen.org/tools/nest/branch/mod_python@975 e0e46c49-be69-4f5a-ad62-21024a331aea