nest: optimize dorf data fetch, re #111
Since the koordinates in the dorf table are both
in one cell, we can't crop these easily.
So we fetch like 10 k rows every time,
even for the imported map.

I replaced the row[0] regex with split()
and cached config.is_kraehe (another regex)

This is an 0.3 second improvement (currently 1/2 of total)


git-svn-id: http://svn.kraehen.org/tools/nest/branch/mod_python@977 e0e46c49-be69-4f5a-ad62-21024a331aea