nest: add crop clause for dorf fetch, close #111
The dorf table is somewhat broken because the x and y
are both in the koords column as "x,y".
However, we still have to crop somehow,
otherwise we always fetch all rows!

We use multiple LIKE clauses now.
One for every x in xmin, xmax and the same for y.
That is not the same as <= and >=, but better than nothing.

That is possibly the last big thing we can do
and probably god enough for now.
The import map only takes 0.1 to 0.2 seconds.
(rather than 0.4-0.5)


git-svn-id: http://svn.kraehen.org/tools/nest/branch/mod_python@979 e0e46c49-be69-4f5a-ad62-21024a331aea