nest: small_map: fetch armeen last, re #112
Armeen were probably changed at the same time the small map was fetched.
We should fetch them as late as possible.

This might still not help and we have to change kraehenauge
to request the map later.

Before some optimizations the dorf fetch took very long
so the Armee data was probably already updated afterwards.


git-svn-id: http://svn.kraehen.org/tools/nest/trunk@986 e0e46c49-be69-4f5a-ad62-21024a331aea